added services for foodSearchHomePage, history (calorie tab only), and gathering tokens, also added loading for home page
